Separate application interface from daemons.
