+++ src/beamlime/applications/daemons.py
@@ -1,16 +1,131 @@
 # SPDX-License-Identifier: BSD-3-Clause
 # Copyright (c) 2024 Scipp contributors (https://github.com/scipp)
-from abc import ABC, abstractmethod, abstractproperty
-from typing import Any, List
+import asyncio
+from contextlib import contextmanager
+from typing import Any, Awaitable, Callable, Coroutine, Generator, List, Optional
 
-from ..constructors import ProviderGroup, SingletonProvider
-from ..logging import BeamlimeLogger
-from ..logging.mixins import LogMixin
 from ._parameters import ChunkSize, DataFeedingSpeed
 from ._random_data_providers import RandomEvents
 from ._workflow import Events
-from .base import BaseDaemon, MessageRouter
-from .handlers import DataReductionHandler, HistogramUpdated, PlotSaver, RawDataSent
+from .base import BaseDaemon, BeamlimeMessage
+from .handlers import RawDataSent
+
+
+class MessageRouter(BaseDaemon):
+    """A message router that routes messages to handlers."""
+
+    message_pipe: List[BeamlimeMessage]
+
+    class StopRouting(BeamlimeMessage):
+        ...
+
+    def __init__(self):
+        self.handlers: dict[type, List[Callable[[BeamlimeMessage], Any]]] = dict()
+        self.awaitable_handlers: dict[
+            type, List[Callable[[BeamlimeMessage], Awaitable[Any]]]
+        ] = dict()
+        self.message_pipe = list()
+        self._break_routing_loop = False  # Break the routing loop flag
+
+    def break_routing_loop(self, message: Optional[BeamlimeMessage] = None) -> None:
+        if not isinstance(message, self.StopRouting):
+            raise TypeError(
+                f"Expected message of type {self.StopRouting}, got {type(message)}."
+            )
+        self._break_routing_loop = True
+
+    @contextmanager
+    def _handler_wrapper(
+        self, handler: Callable[..., Any], message: BeamlimeMessage
+    ) -> Generator[None, None, None]:
+        import warnings
+
+        try:
+            self.debug(f"Routing event {type(message)} to handler {handler}...")
+            yield
+        except Exception as e:
+            warnings.warn(f"Failed to handle event {type(message)}: {e}", stacklevel=2)
+        else:
+            self.debug(f"Routing event {type(message)} to handler {handler} done.")
+
+    def _register(
+        self,
+        *,
+        handler_list: dict[
+            type[BeamlimeMessage], List[Callable[[BeamlimeMessage], Any]]
+        ],
+        event_tp: type[BeamlimeMessage],
+        handler: Callable[[BeamlimeMessage], Any],
+    ):
+        if event_tp in handler_list:
+            handler_list[event_tp].append(handler)
+        else:
+            handler_list[event_tp] = [handler]
+
+    def register_awaitable_handler(
+        self, event_tp, handler: Callable[[BeamlimeMessage], Coroutine[Any, Any, Any]]
+    ):
+        self._register(
+            handler_list=self.awaitable_handlers, event_tp=event_tp, handler=handler
+        )
+
+    def register_handler(self, event_tp, handler: Callable[[BeamlimeMessage], Any]):
+        self._register(handler_list=self.handlers, event_tp=event_tp, handler=handler)
+
+    def _check_result(self, result: Any) -> None:
+        """Append or extend ``result`` to ``self.message_pipe``.
+
+        It filters out non-BeamlimeMessage objects from ``result``.
+        """
+        if isinstance(result, BeamlimeMessage):
+            self.message_pipe.append(result)
+        elif isinstance(result, tuple):
+            msgs = (_msg for _msg in result if isinstance(_msg, BeamlimeMessage))
+            self.message_pipe.extend(msgs)
+
+    async def route(self, message: BeamlimeMessage) -> None:
+        # Synchronous handlers
+        if handlers := self.handlers.get(type(message), []):
+            for handler in handlers:
+                await asyncio.sleep(0)  # Let others use the event loop.
+                with self._handler_wrapper(handler, message):
+                    self._check_result(handler(message))
+
+        # Asynchronous handlers
+        if awaitable_handlers := self.awaitable_handlers.get(type(message), []):
+            for handler in awaitable_handlers:
+                with self._handler_wrapper(handler, message):
+                    self._check_result(await handler(message))
+
+        # No handlers
+        if not (handlers or awaitable_handlers):
+            import warnings
+
+            warnings.warn(
+                f"No handler for event {type(message)}. Ignoring...", stacklevel=2
+            )
+
+    async def run(self) -> None:
+        """Message router daemon."""
+        data_monitor = self.data_pipe_monitor(
+            self.message_pipe,
+            timeout=0.1,
+            interval=0.1,
+        )
+
+        while not self._break_routing_loop:
+            if await data_monitor():
+                message = self.message_pipe.pop(0)
+                await self.route(message)
+
+        self.debug("Breaking routing loop. Routing the rest of the message...")
+        while self.message_pipe:
+            await self.route(self.message_pipe.pop(0))
+        self.debug("Routing the rest of the message done.")
+
+    async def send_message_async(self, message: BeamlimeMessage) -> None:
+        self.message_pipe.append(message)
+        await asyncio.sleep(0)
 
 
 class DataStreamSimulator(BaseDaemon):
@@ -56,150 +171,3 @@ class DataStreamSimulator(BaseDaemon):
         )
 
         self.info("Data streaming finished...")
-
-
-class ApplicationInterface(LogMixin, ABC):
-    """Base class for applications."""
-
-    logger: BeamlimeLogger
-
-    def __init__(self) -> None:
-        import asyncio
-
-        self.loop: asyncio.AbstractEventLoop
-        self.tasks: List[asyncio.Task]
-        super().__init__()
-
-    @abstractmethod
-    def register_handling_methods(self):
-        """Register handlers to the application."""
-        ...
-
-    @abstractproperty
-    def daemons(self) -> tuple[BaseDaemon, ...]:
-        """Collection of daemons that are running in the application."""
-        ...
-
-    def run(self):
-        """
-        Collect all coroutines of daemons and schedule them into the event loop.
-
-        Notes
-        -----
-        **Debugging log while running async daemons under various circumstances.**
-
-        - ``asyncio.get_event_loop`` vs ``asyncio.new_event_loop``
-        1. ``asyncio.get_event_loop``
-        ``get_event_loop`` will always return the current event loop.
-        If there is no event loop set in the thread, it will create a new one
-        and set it as a current event loop of the thread, and return the loop.
-        Many of ``asyncio`` free functions internally use ``get_event_loop``,
-        i.e. ``asyncio.create_task``.
-
-        **Things to be considered while using ``asyncio.get_event_loop``.
-          - ``asyncio.create_task`` does not guarantee
-            whether the current loop is/will be alive until the task is done.
-            You may use ``run_until_complete`` to make sure the loop is not closed
-            until the task is finished.
-            When you need to throw multiple async calls to the loop,
-            use ``asyncio.gather`` to merge all the tasks like in this method.
-          - ``close`` or ``stop`` might accidentally destroy/interrupt
-            other tasks running in the same event loop.
-            i.e. You can accidentally destroy the main event loop of a jupyter kernel.
-          - [1]``RuntimeError`` if there has been an event loop set in the
-            thread object before but it is now removed.
-
-        2. ``asyncio.new_event_loop``
-        ``asyncio.new_event_loop`` will always return the new event loop,
-        but it is not set it as a current loop of the thread automatically.
-
-        However, sometimes it is automatically handled within the thread,
-        and it caused errors which was hard to be debugged under ``pytest`` session.
-        For example,
-        - The new event loop was not closed properly as it is destroyed.
-        - The new event loop was never started until it is destroyed.
-        ``Traceback`` of ``pytest`` did not show
-        where exactly the error is from in those cases.
-        It was resolved by using ``get_event_loop``,
-        or manually closing the event loop at the end of the test.
-
-        **When to use ``asyncio.new_event_loop``.**
-          - ``asyncio.get_event_loop`` raises ``RuntimeError``[1]
-          - Multi-threads
-
-        Please note that the loop object might need to be ``close``ed manually.
-        """
-        import asyncio
-        import time
-
-        from beamlime.core.schedulers import temporary_event_loop
-
-        self.register_handling_methods()
-        self.info('Start running %s...', self.__class__.__qualname__)
-        start = time.time()
-        with temporary_event_loop() as loop:
-            self.loop = loop
-            daemon_coroutines = [daemon.run() for daemon in self.daemons]
-            self.tasks = [loop.create_task(coro) for coro in daemon_coroutines]
-            if not loop.is_running():
-                loop.run_until_complete(asyncio.gather(*self.tasks))
-        self.info('Finished running %s...', time.time() - start)
-
-
-class DataReductionApp(ApplicationInterface):
-    """Data reduction daemon that handles all part of the data reduction process."""
-
-    # Daemons
-    data_stream_listener: DataStreamSimulator
-    message_router: MessageRouter
-    # Handlers
-    data_reduction_handler: DataReductionHandler
-    plot_handler: PlotSaver
-
-    def register_handling_methods(self) -> None:
-        # Message Router
-        self.message_router.register_handler(
-            self.message_router.StopRouting, self.message_router.break_routing_loop
-        )
-        # Data Reduction Handler
-        self.message_router.register_awaitable_handler(
-            RawDataSent, self.data_reduction_handler.process_message
-        )
-        # Plot Handler
-        self.message_router.register_awaitable_handler(
-            HistogramUpdated, self.plot_handler.save_histogram
-        )
-
-    @property
-    def daemons(self) -> tuple[BaseDaemon, ...]:
-        return (
-            self.data_stream_listener,
-            self.message_router,
-        )
-
-    @staticmethod
-    def collect_default_providers() -> ProviderGroup:
-        """Helper method to collect all default providers for this prototype."""
-        from beamlime.applications._parameters import collect_default_param_providers
-        from beamlime.applications._random_data_providers import random_data_providers
-        from beamlime.applications._workflow import provide_pipeline
-        from beamlime.applications.handlers import random_image_path
-        from beamlime.constructors.providers import merge as merge_providers
-        from beamlime.logging.providers import log_providers
-
-        app_providers = ProviderGroup(
-            SingletonProvider(DataReductionApp),
-            DataStreamSimulator,
-            DataReductionHandler,
-            PlotSaver,
-            provide_pipeline,
-            random_image_path,
-        )
-        app_providers[MessageRouter] = SingletonProvider(MessageRouter)
-
-        return merge_providers(
-            collect_default_param_providers(),
-            random_data_providers,
-            app_providers,
-            log_providers,
-        )
